Move layout for multiple choice field
