+++ gnd/src/main/java/com/google/android/gnd/ui/common/BindingAdapters.java
@@ -31,7 +31,7 @@ import com.google.android.gnd.databinding.MultipleChoiceInputFieldBinding;
 import com.google.android.gnd.databinding.TextInputFieldBinding;
 import com.google.android.gnd.model.form.Field;
 import com.google.android.gnd.model.observation.Response;
-import com.google.android.gnd.ui.editobservation.MultipleChoiceFieldLayout;
+import com.google.android.gnd.ui.editobservation.field.MultipleChoiceFieldLayout;
 import com.google.android.material.textfield.TextInputEditText;
 import com.squareup.picasso.Picasso;
 import java8.util.function.Consumer;
